Added auth and code layout changes

+++ src/com/datadynamic/server/MovieServiceImpl.java
@@ -10,6 +10,9 @@ import org.apache.commons.dbutils.handlers.BeanHandler;
 import org.apache.commons.dbutils.handlers.BeanListHandler;
 
 import com.datadynamic.client.remoteservices.MovieService;
+import com.datadynamic.server.utils.AuthenticationUtils;
+import com.datadynamic.server.utils.DB;
+import com.datadynamic.server.utils.ServerLog;
 import com.datadynamic.shared.Labels;
 import com.datadynamic.shared.pojos.*;
 import com.google.gwt.user.server.rpc.RemoteServiceServlet;
@@ -18,11 +21,12 @@ import com.google.gwt.user.server.rpc.RemoteServiceServlet;
  * The server-side implementation of the RPC service.
  */
 @SuppressWarnings("serial")
-public class MovieServiceImpl extends RemoteServiceServlet implements MovieService {	
-	
-	
+public class MovieServiceImpl extends RemoteServiceServlet implements MovieService {		
 	@Override
 	public ArrayList<Movie> getMovieCollection() throws IllegalArgumentException {
+		if(!AuthenticationUtils.IsAuthenticated(this.getThreadLocalRequest())) {
+			return null;
+		}
 		QueryRunner run = new QueryRunner();
 		ResultSetHandler<List<Movie>> h = new BeanListHandler<Movie>(Movie.class);
 		List<Movie> movies = null;
@@ -30,13 +34,16 @@ public class MovieServiceImpl extends RemoteServiceServlet implements MovieServi
 			movies = run.query(DB.getConnection(), "SELECT * FROM Movies", h);
 		} catch (SQLException e) {
 			e.printStackTrace();
-			ServerLog.error(e.getMessage());			
+			ServerLog.error(e.getMessage());
 		}
 		return new ArrayList<Movie>(movies);
 	}
 
 	@Override
 	public ActionResponse removeMovie(long ID) {
+		if(!AuthenticationUtils.IsAuthenticated(this.getThreadLocalRequest())) {
+			return null;
+		}	
 		QueryRunner run = new QueryRunner();
 		try {
 			run.update(DB.getConnection(), "delete from MOVIES where ID = ?", ID);
@@ -50,6 +57,9 @@ public class MovieServiceImpl extends RemoteServiceServlet implements MovieServi
 
 	@Override
 	public ActionResponse addMovie(String name, String genre) {
+		if(!AuthenticationUtils.IsAuthenticated(this.getThreadLocalRequest())) {
+			return null;
+		}
 		QueryRunner run = new QueryRunner();
 		ResultSetHandler<Movie> h = new BeanHandler<Movie>(Movie.class);
 		try {
